SDIT-2832 Refactor - rename to TAP and split out service for each entity
