+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonertonomisupdate/movements/ExternalMovementsReconciliationResource.kt
@@ -5,6 +5,10 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RestController
+import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.taps.MismatchPrisonerTaps
+import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.taps.MismatchPrisonerTapsSummary
+import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.taps.TemporaryAbsencesActivePrisonersReconciliationService
+import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.taps.TemporaryAbsencesAllPrisonersReconciliationService
 
 @PreAuthorize("hasAnyRole('PRISONER_TO_NOMIS__UPDATE__RW', 'ROLE_PRISONER_FROM_NOMIS__REPAIR_MOVEMENTS__RW')")
 @RestController
