+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonertonomisupdate/movements/ExternalMovementsToNomisIntTest.kt
@@ -14,7 +14,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.untilAsserted
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
@@ -31,13 +30,11 @@ import software.amazon.awssdk.services.sns.model.MessageAttributeValue
 import software.amazon.awssdk.services.sns.model.PublishRequest
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.integration.SqsIntegrationTestBase
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.ExternalMovementsDpsApiExtension.Companion.dpsExternalMovementsServer
-import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.ExternalMovementsNomisApiMockServer.Companion.createTemporaryAbsenceResponse
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.ExternalMovementsNomisApiMockServer.Companion.upsertScheduledTemporaryAbsenceResponse
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.ExternalMovementsNomisApiMockServer.Companion.upsertTemporaryAbsenceApplicationResponse
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.movements.model.Location
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.nomismappings.model.DuplicateErrorContentObject
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.nomismappings.model.DuplicateMappingErrorResponse
-import uk.gov.justice.digital.hmpps.prisonertonomisupdate.nomismappings.model.ExternalMovementSyncMappingDto
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.nomismappings.model.ScheduledMovementSyncMappingDto
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.nomismappings.model.TemporaryAbsenceApplicationSyncMappingDto
 import uk.gov.justice.digital.hmpps.prisonertonomisupdate.wiremock.withRequestBodyJsonPath
@@ -1353,407 +1350,6 @@ class ExternalMovementsToNomisIntTest : SqsIntegrationTestBase() {
     }
   }
 
-  @Nested
-  @DisplayName("external-movements-api.temporary-absence-external-movement-out.created")
-  inner class TemporaryAbsenceExternalMovementOutCreated {
-    private val prisonerNumber = "A1234BC"
-    private val dpsExternalMovementId = UUID.randomUUID()
-    private val dpsScheduledMovementId = UUID.randomUUID()
-    private val dpsMovementApplicationId = UUID.randomUUID()
-    private val nomisBookingId = 12345L
-    private val nomisMovementSeq = 2
-    private val nomisMovementApplicationId = 543L
-    private val nomisEventId = 54321L
-
-    @Nested
-    @DisplayName("when NOMIS is the origin of an external movement create")
-    inner class WhenNomisCreated {
-
-      @BeforeEach
-      fun setUp() {
-        publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId = dpsExternalMovementId, dpsScheduledMovementId = dpsScheduledMovementId, prisonerNumber = prisonerNumber, dpsAuthorisationId = dpsMovementApplicationId, source = "NOMIS")
-        waitForAnyProcessingToComplete("temporary-absence-external-movement-create-ignored")
-      }
-
-      @Test
-      fun `will send telemetry event showing the ignore`() {
-        verify(telemetryClient).trackEvent(
-          eq("temporary-absence-external-movement-create-ignored"),
-          any(),
-          isNull(),
-        )
-      }
-    }
-
-    @Nested
-    @DisplayName("when DPS is the origin of an external movement create")
-    inner class WhenDpsCreated {
-      @Nested
-      @DisplayName("when all goes ok")
-      inner class HappyPath {
-
-        @BeforeEach
-        fun setUp() {
-          mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-          mappingApi.stubGetTemporaryAbsenceScheduledMovementMapping(dpsId = dpsScheduledMovementId, nomisEventId = nomisEventId)
-          mappingApi.stubGetTemporaryAbsenceApplicationMapping(dpsId = dpsMovementApplicationId, nomisMovementApplicationId = nomisMovementApplicationId)
-// TODO         dpsApi.stubGetTemporaryAbsenceExternalMovement(dpsId, temporaryAbsenceExternalMovement())
-          nomisApi.stubCreateTemporaryAbsence(prisonerNumber, createTemporaryAbsenceResponse(nomisBookingId, nomisMovementSeq))
-          mappingApi.stubCreateExternalMovementMapping()
-          publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, dpsScheduledMovementId, prisonerNumber, dpsMovementApplicationId, "DPS")
-          waitForAnyProcessingToComplete()
-        }
-
-        @Test
-        fun `will send telemetry event showing the create`() {
-          verify(telemetryClient).trackEvent(
-            eq("temporary-absence-external-movement-create-success"),
-            any(),
-            isNull(),
-          )
-        }
-
-        @Test
-        fun `telemetry will contain key facts about the external movement created`() {
-          verify(telemetryClient).trackEvent(
-            eq("temporary-absence-external-movement-create-success"),
-            check {
-              assertThat(it).containsEntry("dpsMovementApplicationId", dpsMovementApplicationId.toString())
-              assertThat(it).containsEntry("dpsScheduledMovementOutId", dpsScheduledMovementId.toString())
-              assertThat(it).containsEntry("dpsExternalMovementId", dpsExternalMovementId.toString())
-              assertThat(it).containsEntry("nomisEventId", nomisEventId.toString())
-              assertThat(it).containsEntry("prisonerNumber", prisonerNumber)
-              assertThat(it).containsEntry("bookingId", "12345")
-              assertThat(it).containsEntry("nomisMovementSeq", "2")
-              assertThat(it).containsEntry("direction", "OUT")
-            },
-            isNull(),
-          )
-        }
-
-        @Test
-        fun `will check if the external movement mapping exists`() {
-          mappingApi.verify(getRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement/dps-id/$dpsExternalMovementId")))
-        }
-
-        @Test
-        fun `will check if the parent scheduled movement mapping exists`() {
-          mappingApi.verify(getRequestedFor(urlEqualTo("/mapping/temporary-absence/application/dps-id/$dpsMovementApplicationId")))
-        }
-
-        @Test
-        fun `will check if the parent application mapping exists`() {
-          mappingApi.verify(getRequestedFor(urlEqualTo("/mapping/temporary-absence/application/dps-id/$dpsMovementApplicationId")))
-        }
-
-        @Test
-        @Disabled("TODO waiting for DPS API")
-        fun `will call back to DPS to get external details`() {
-// TODO         dpsApi.verify(getRequestedFor(urlEqualTo("/sync/temporary-absence-external-movement/$dpsId")))
-        }
-
-        @Test
-        fun `will create the external movement in NOMIS`() {
-          nomisApi.verify(postRequestedFor(urlPathEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-        }
-
-        @Test
-        fun `the created external movement will contain correct details`() {
-          nomisApi.verify(
-            postRequestedFor(anyUrl())
-              .withRequestBodyJsonPath("scheduledTemporaryAbsenceId", nomisEventId)
-              .withRequestBodyJsonPath("movementDate", today.toLocalDate())
-              .withRequestBodyJsonPath("movementReason", "C5")
-              .withRequestBodyJsonPath("arrestAgency", "POL")
-              .withRequestBodyJsonPath("escort", "U")
-              .withRequestBodyJsonPath("escortText", "Temporary absence escort text")
-              .withRequestBodyJsonPath("fromPrison", "LEI")
-              .withRequestBodyJsonPath("toAgency", "HAZLWD")
-              .withRequestBodyJsonPath("commentText", "Temporary absence comment")
-              .withRequestBodyJsonPath("toAddressId", 76543)
-              .withRequestBodyJsonPath("toCity", "765"),
-          )
-        }
-
-        @Test
-        fun `will create a mapping between the NOMIS and DPS ids`() {
-          mappingApi.verify(
-            postRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement"))
-              .withRequestBodyJsonPath("dpsMovementId", "$dpsExternalMovementId")
-              .withRequestBodyJsonPath("nomisMovementSeq", "$nomisMovementSeq")
-              .withRequestBodyJsonPath("prisonerNumber", "A1234BC")
-              .withRequestBodyJsonPath("bookingId", "$nomisBookingId")
-              .withRequestBodyJsonPath("mappingType", "DPS_CREATED"),
-            // TODO check address mappings
-          )
-        }
-      }
-
-      @Nested
-      @DisplayName("when mapping service fails once")
-      inner class MappingFailure {
-
-        @BeforeEach
-        fun setUp() {
-          mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-          mappingApi.stubGetTemporaryAbsenceScheduledMovementMapping(dpsId = dpsScheduledMovementId, nomisEventId = nomisEventId)
-          mappingApi.stubGetTemporaryAbsenceApplicationMapping(dpsId = dpsMovementApplicationId, nomisMovementApplicationId = nomisMovementApplicationId)
-// TODO         dpsApi.stubGetTemporaryAbsenceExternalMovement(dpsId, temporaryAbsenceExternalMovement())
-          nomisApi.stubCreateTemporaryAbsence(prisonerNumber, createTemporaryAbsenceResponse(nomisBookingId, nomisMovementSeq))
-          mappingApi.stubCreateExternalMovementMappingFailureFollowedBySuccess()
-          publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, dpsScheduledMovementId, prisonerNumber, dpsMovementApplicationId, "DPS")
-        }
-
-        @Test
-        fun `will send telemetry for initial failure`() {
-          await untilAsserted {
-            verify(telemetryClient).trackEvent(
-              eq("temporary-absence-external-movement-mapping-create-error"),
-              any(),
-              isNull(),
-            )
-          }
-        }
-
-        @Test
-        fun `will eventually send telemetry for success`() {
-          await untilAsserted {
-            verify(telemetryClient).trackEvent(
-              eq("temporary-absence-external-movement-create-success"),
-              any(),
-              isNull(),
-            )
-          }
-        }
-
-        @Test
-        fun `will create the external movement in NOMIS once`() {
-          await untilAsserted {
-            verify(telemetryClient).trackEvent(
-              eq("temporary-absence-external-movement-create-success"),
-              any(),
-              isNull(),
-            )
-            nomisApi.verify(1, postRequestedFor(urlEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-          }
-        }
-      }
-
-      @Nested
-      @DisplayName("when mapping service detects a duplicate mapping")
-      inner class DuplicateMappingFailure {
-
-        @BeforeEach
-        fun setUp() {
-          mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-          mappingApi.stubGetTemporaryAbsenceScheduledMovementMapping(dpsId = dpsScheduledMovementId, nomisEventId = nomisEventId)
-          mappingApi.stubGetTemporaryAbsenceApplicationMapping(dpsId = dpsMovementApplicationId, nomisMovementApplicationId = nomisMovementApplicationId)
-// TODO         dpsApi.stubGetTemporaryAbsenceExternalMovement(dpsId, temporaryAbsenceExternalMovement())
-          nomisApi.stubCreateTemporaryAbsence(prisonerNumber, createTemporaryAbsenceResponse(nomisBookingId, nomisMovementSeq))
-          mappingApi.stubCreateExternalMovementMappingConflict(
-            error = DuplicateMappingErrorResponse(
-              moreInfo = DuplicateErrorContentObject(
-                existing = ExternalMovementSyncMappingDto(
-                  prisonerNumber = "A1234BC",
-                  bookingId = nomisBookingId,
-                  nomisMovementSeq = nomisMovementSeq,
-                  dpsMovementId = dpsExternalMovementId,
-                  mappingType = ExternalMovementSyncMappingDto.MappingType.NOMIS_CREATED,
-                  // TODO add address mapping details
-                  nomisAddressId = 0,
-                  nomisAddressOwnerClass = "",
-                  dpsAddressText = "",
-                ),
-                duplicate = ExternalMovementSyncMappingDto(
-                  prisonerNumber = "A1234BC",
-                  bookingId = nomisBookingId,
-                  nomisMovementSeq = nomisMovementSeq + 1,
-                  dpsMovementId = dpsExternalMovementId,
-                  mappingType = ExternalMovementSyncMappingDto.MappingType.NOMIS_CREATED,
-                  // TODO add address mapping details
-                  nomisAddressId = 0,
-                  nomisAddressOwnerClass = "",
-                  dpsAddressText = "",
-                ),
-              ),
-              errorCode = 1409,
-              status = DuplicateMappingErrorResponse.Status._409_CONFLICT,
-              userMessage = "Duplicate mapping",
-            ),
-          )
-
-          publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, dpsScheduledMovementId, prisonerNumber, dpsMovementApplicationId, "DPS")
-        }
-
-        @Test
-        fun `will send telemetry for duplicate`() {
-          await untilAsserted {
-            verify(telemetryClient).trackEvent(
-              eq("to-nomis-synch-temporary-absence-external-movement-duplicate"),
-              any(),
-              isNull(),
-            )
-          }
-        }
-
-        @Test
-        fun `will create the person in NOMIS once`() {
-          await untilAsserted {
-            verify(telemetryClient).trackEvent(
-              eq("to-nomis-synch-temporary-absence-external-movement-duplicate"),
-              any(),
-              isNull(),
-            )
-            nomisApi.verify(1, postRequestedFor(urlEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-          }
-        }
-      }
-
-      @Nested
-      @DisplayName("when parent application mapping does not exist")
-      inner class ParentApplicationMappingNotFound {
-
-        @BeforeEach
-        fun setUp() {
-          mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-          mappingApi.stubGetTemporaryAbsenceApplicationMapping(dpsId = dpsMovementApplicationId, status = NOT_FOUND)
-
-          publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, dpsScheduledMovementId, prisonerNumber, dpsMovementApplicationId, "DPS")
-          waitForAnyProcessingToComplete("temporary-absence-external-movement-create-error")
-        }
-
-        @Test
-        fun `will not create the application in NOMIS`() {
-          nomisApi.verify(count = 0, postRequestedFor(urlPathEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-        }
-
-        @Test
-        fun `will not create a mapping between the NOMIS and DPS ids`() {
-          mappingApi.verify(
-            count = 0,
-            postRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement")),
-          )
-        }
-      }
-
-      @Nested
-      @DisplayName("when parent scheduled movement mapping does not exist")
-      inner class ParentScheduledMovementMappingNotFound {
-
-        @BeforeEach
-        fun setUp() {
-          mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-          mappingApi.stubGetTemporaryAbsenceApplicationMapping(dpsId = dpsMovementApplicationId, nomisMovementApplicationId = nomisMovementApplicationId)
-          mappingApi.stubGetTemporaryAbsenceScheduledMovementMapping(dpsId = dpsScheduledMovementId, status = NOT_FOUND)
-
-          publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, dpsScheduledMovementId, prisonerNumber, dpsMovementApplicationId, "DPS")
-          waitForAnyProcessingToComplete("temporary-absence-external-movement-create-error")
-        }
-
-        @Test
-        fun `will not create the application in NOMIS`() {
-          nomisApi.verify(count = 0, postRequestedFor(urlPathEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-        }
-
-        @Test
-        fun `will not create a mapping between the NOMIS and DPS ids`() {
-          mappingApi.verify(
-            count = 0,
-            postRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement")),
-          )
-        }
-      }
-    }
-
-    @Nested
-    @DisplayName("Unscheduled movement, when all goes ok")
-    inner class UnscheduledHappyPath {
-
-      @BeforeEach
-      fun setUp() {
-        mappingApi.stubGetTemporaryAbsenceExternalMovementMapping(dpsId = dpsExternalMovementId, status = NOT_FOUND)
-// TODO         dpsApi.stubGetTemporaryAbsenceExternalMovement(dpsId, temporaryAbsenceExternalMovement())
-        nomisApi.stubCreateTemporaryAbsence(prisonerNumber, createTemporaryAbsenceResponse(nomisBookingId, nomisMovementSeq))
-        mappingApi.stubCreateExternalMovementMapping()
-        publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId, null, prisonerNumber, null, "DPS")
-        waitForAnyProcessingToComplete()
-      }
-
-      @Test
-      fun `will send telemetry event showing the create`() {
-        verify(telemetryClient).trackEvent(
-          eq("temporary-absence-external-movement-create-success"),
-          any(),
-          isNull(),
-        )
-      }
-
-      @Test
-      fun `telemetry will contain key facts about the external movement created`() {
-        verify(telemetryClient).trackEvent(
-          eq("temporary-absence-external-movement-create-success"),
-          check {
-            assertThat(it["dpsMovementApplicationId"]).isNull()
-            assertThat(it["dpsScheduledMovementId"]).isNull()
-            assertThat(it).containsEntry("dpsExternalMovementId", dpsExternalMovementId.toString())
-            assertThat(it).containsEntry("prisonerNumber", prisonerNumber)
-            assertThat(it).containsEntry("bookingId", "12345")
-            assertThat(it).containsEntry("nomisMovementSeq", "2")
-            assertThat(it).containsEntry("direction", "OUT")
-          },
-          isNull(),
-        )
-      }
-
-      @Test
-      fun `will check if the external movement mapping exists`() {
-        mappingApi.verify(getRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement/dps-id/$dpsExternalMovementId")))
-      }
-
-      @Test
-      @Disabled("TODO waiting for DPS API")
-      fun `will call back to DPS to get external details`() {
-// TODO         dpsApi.verify(getRequestedFor(urlEqualTo("/sync/temporary-absence-external-movement/$dpsId")))
-      }
-
-      @Test
-      fun `will create the external movement in NOMIS`() {
-        nomisApi.verify(postRequestedFor(urlPathEqualTo("/movements/A1234BC/temporary-absences/temporary-absence")))
-      }
-
-      @Test
-      fun `the created external movement will contain correct details`() {
-        nomisApi.verify(
-          postRequestedFor(anyUrl())
-            .withRequestBodyJsonPath("scheduledTemporaryAbsenceId", absent())
-            .withRequestBodyJsonPath("movementDate", today.toLocalDate())
-            .withRequestBodyJsonPath("movementReason", "C5")
-            .withRequestBodyJsonPath("arrestAgency", "POL")
-            .withRequestBodyJsonPath("escort", "U")
-            .withRequestBodyJsonPath("escortText", "Temporary absence escort text")
-            .withRequestBodyJsonPath("fromPrison", "LEI")
-            .withRequestBodyJsonPath("toAgency", "HAZLWD")
-            .withRequestBodyJsonPath("commentText", "Temporary absence comment")
-            .withRequestBodyJsonPath("toAddressId", 76543)
-            .withRequestBodyJsonPath("toCity", "765"),
-        )
-      }
-
-      @Test
-      fun `will create a mapping between the NOMIS and DPS ids`() {
-        mappingApi.verify(
-          postRequestedFor(urlEqualTo("/mapping/temporary-absence/external-movement"))
-            .withRequestBodyJsonPath("dpsMovementId", "$dpsExternalMovementId")
-            .withRequestBodyJsonPath("nomisMovementSeq", "$nomisMovementSeq")
-            .withRequestBodyJsonPath("prisonerNumber", "A1234BC")
-            .withRequestBodyJsonPath("bookingId", "$nomisBookingId")
-            .withRequestBodyJsonPath("mappingType", "DPS_CREATED"),
-          // TODO check address mappings
-        )
-      }
-    }
-  }
-
   private fun publishAuthorisationDomainEvent(dpsAuthorisationId: UUID, prisonerNumber: String, source: String = "DPS", eventType: String = "person.temporary-absence-authorisation.approved") {
     with(eventType) {
       publishDomainEvent(eventType = this, payload = messagePayload(eventType = this, id = dpsAuthorisationId, prisonerNumber = prisonerNumber, source = source))
@@ -1794,24 +1390,6 @@ class ExternalMovementsToNomisIntTest : SqsIntegrationTestBase() {
     }
     """
 
-  private fun publishTemporaryAbsenceExternalMovementOutDomainEvent(dpsExternalMovementId: UUID, dpsScheduledMovementId: UUID? = null, prisonerNumber: String, dpsAuthorisationId: UUID? = null, source: String = "DPS") {
-    with("external-movements-api.temporary-absence-external-movement-out.created") {
-      publishDomainEvent(
-        eventType = this,
-        payload = messagePayloadExternalMovements(eventType = this, dpsAuthorisationId = dpsAuthorisationId, dpsScheduledMovementOutId = dpsScheduledMovementId, dpsExternalMovementOutId = dpsExternalMovementId, prisonerNumber = prisonerNumber, source = source),
-      )
-    }
-  }
-
-  private fun publishTemporaryAbsenceExternalMovementInDomainEvent(dpsExternalMovementId: UUID, dpsScheduledMovementId: UUID? = null, prisonerNumber: String, dpsAuthorisationId: UUID? = null, source: String = "DPS") {
-    with("external-movements-api.temporary-absence-external-movement-in.created") {
-      publishDomainEvent(
-        eventType = this,
-        payload = messagePayloadExternalMovements(eventType = this, dpsAuthorisationId = dpsAuthorisationId, dpsScheduledMovementInId = dpsScheduledMovementId, dpsExternalMovementInId = dpsExternalMovementId, prisonerNumber = prisonerNumber, source = source),
-      )
-    }
-  }
-
   private fun publishDomainEvent(
     eventType: String,
     payload: String,
@@ -1828,38 +1406,3 @@ class ExternalMovementsToNomisIntTest : SqsIntegrationTestBase() {
     ).get()
   }
 }
-
-fun messagePayloadExternalMovements(
-  eventType: String,
-  prisonerNumber: String,
-  dpsAuthorisationId: UUID? = null,
-  dpsScheduledMovementOutId: UUID? = null,
-  dpsScheduledMovementInId: UUID? = null,
-  dpsOccurrenceId: UUID? = null,
-  dpsExternalMovementOutId: UUID? = null,
-  dpsExternalMovementInId: UUID? = null,
-  source: String,
-) = //language=JSON
-  """
-    {
-      "description":"Some event", 
-      "eventType":"$eventType", 
-      "additionalInformation": {
-        ${dpsAuthorisationId?.let { """"authorisationId": "$dpsAuthorisationId",""" } ?: ""}
-        ${dpsOccurrenceId?.let { """"occurrenceId": "$dpsOccurrenceId",""" } ?: ""}
-        ${dpsScheduledMovementOutId?.let { """"scheduledMovementOutId": "$dpsScheduledMovementOutId",""" } ?: ""}
-        ${dpsScheduledMovementInId?.let { """"scheduledMovementInId": "$dpsScheduledMovementInId",""" } ?: ""}
-        ${dpsExternalMovementOutId?.let { """"externalMovementOutId": "$dpsExternalMovementOutId",""" } ?: ""}
-        ${dpsExternalMovementInId?.let { """"externalMovementInId": "$dpsExternalMovementInId",""" } ?: ""}
-        "source": "$source"
-      },
-      "personReference": {
-        "identifiers": [
-          {
-            "type": "NOMS",
-            "value": "$prisonerNumber"
-          }
-        ]
-      }
-    }
-    """
